Reworked Rotations
Rotations have been "fixed", and camera movement is now relative to theta as well (pressing W moves up if theta is 0, but sideways if theta is 90 degrees for example). 

However, getEnvCoordinates() still needs to be fixed up, as the code does not work for any reference point not in the first quadrant.